requirements.txt
flask
pytest
gunicorn
# SQLAlchemy
flask-sqlalchemy
flasgger
Flask-Migrate
flask-admin
python-decouple